+++ ServerREST/src/apistate.cpp
@@ -43,10 +43,10 @@ void ApiState::RoomId_GET(Context *c, const QString &roomid)
     if (!serve.isValid(roomId)) return;
     Room* r = serve.getRoom(roomId);
     QString state = r->getState();
-    int currentTemp = r->currentTemp;
-    int targetTemp = r->targetTemp;
-    float fee = r->fee;
-    float feerate = r->feerate;
+    double currentTemp = r->currentTemp;
+    double targetTemp = r->targetTemp;
+    double fee = r->fee;
+    double feerate = r->feerate;
     int fanspeed = r->fanspeed;
     int dur = r->serveTime;
